chore(#48): Renamed `common` to `helpers`
