+++ workspaces/backend/gql/src/helpers/asset.ts
