+++ workspaces/backend/gql/src/helpers/image.ts
